doc: update README.md

+++ Lchika/led_on.js
@@ -5,12 +5,27 @@ var gpio3 = dir + '/gpio3';
 
 fs.writeFileSync(dir + '/export', 3); 
 fs.writeFileSync(gpio3 + '/direction', 'out');
-// fs.writeFileSync(gpio3 + '/value', 1);
 
 var cnt = 0;
 
-setInterval(function(){
+function isExistFile(file) {
+  try {
+    fs.statSync(file);
+    return true
+  } catch(err) {
+    if(err.code === 'ENOENT') return false
+  }
+}
+
+var intervalFn = setInterval(function(){
   cnt++;
-  fs.writeFileSync(gpio3 + '/value', cnt % 2);
-  // fs.writeFileSync(dir + '/unexport', 3);
+  if (isExistFile(gpio3)) {
+    fs.writeFileSync(gpio3 + '/value', cnt % 2);
+  }
+  if (cnt === 10) {
+    if (isExistFile(gpio3)) {
+      fs.writeFileSync(dir + '/unexport', 3);
+    }
+    clearInterval(intervalFn);
+  }
 }, 500);
